feat: 맵 스크린 파일 구조 변경, 친구 관리 탭 추가
